+++ packages/dodo-core-attributes/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flightlesslabs/dodo-core-attributes
 
+## 0.2.0
+
+### Minor Changes
+
+- 728fdb5: ComponentAffixPlacement added
+
 ## 0.1.1
 
 ### Patch Changes